// Web/RMA-Web-Shell/projects/clientcare/src/app/product-manager/models/benefit-compensation-amount.ts
export class BenefitCompensationAmount {

  benefitCompensationAmountId: number;
  startDate: Date;
  endDate: Date;
  minCompensationAmount: number;
  maxCompensationAmount: number;
  excessAmount: number;
  benefitId: number;

  isDeleted: boolean;
  createdBy: string;
  modifiedBy: string;
  createdDate: Date;
  modifiedDate: Date;
}